requirements.txt
tkcalendar==1.6.1
openpyxl==3.1.2
reportlab==4.1.0
Pillow==10.3.0
